template_build/pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "tme-template"
version = "1.0.0"
description = "Generator for The Mathematics Educator digital Word template"
requires-python = ">=3.11"
dependencies = [
  "python-docx>=0.8.11",
  "lxml>=4.9",
  "Pillow>=10.0",
  "opencv-python-headless>=4.8",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]
